requirements.txt
python-docx
pdf2image
